requirements.txt
# Core Framework
streamlit>=1.51.0
python-dotenv>=1.2.0

# LlamaIndex Core
llama-index>=0.14.0
llama-index-core>=0.14.0
llama-index-llms-openai>=0.6.0
llama-index-embeddings-huggingface>=0.6.0
llama-index-retrievers-bm25>=0.6.0
llama-index-vector-stores-pinecone>=0.7.0

# Vector Database
pinecone>=7.0.0

# ML & NLP
sentence-transformers>=5.0.0
torch>=2.0.0

# LLM
openai>=2.0.0

# Utilities
pypdf>=6.0.0
rank-bm25>=0.2.0
